+++ Task2/apz-pzpi-21-1-pershyna-anastasiia-task2/GreenGuard/Models/Database/Plant.cs
@@ -18,13 +18,13 @@ namespace GreenGuard.Models.Database
         public required string PlantLocation { get; set; }
 
         [Column("Temp")]
-        public float? Temp { get; set; }
+        public double? Temp { get; set; }
 
         [Column("Humidity")]
-        public float? Humidity { get; set; }
+        public double? Humidity { get; set; }
 
         [Column("Light")]
-        public float? Light { get; set; }
+        public double? Light { get; set; }
 
         [Column("Additional_info")]
         public string? AdditionalInfo { get; set; }
